+++ port/platform/stm32cube/src/u_port_gpio.c
@@ -137,7 +137,7 @@ int32_t uPortGpioSet(int32_t pin, int32_t level)
     uPortPrivateGpioEnableClock(pin);
 
     HAL_GPIO_WritePin(pUPortPrivateGpioGetReg(pin),
-                      1U << U_PORT_STM32F4_GPIO_PIN(pin),
+                      (uint16_t) (1U << U_PORT_STM32F4_GPIO_PIN(pin)),
                       level);
 
     return (int32_t) U_ERROR_COMMON_SUCCESS;
@@ -150,7 +150,7 @@ int32_t uPortGpioGet(int32_t pin)
     uPortPrivateGpioEnableClock(pin);
 
     return HAL_GPIO_ReadPin(pUPortPrivateGpioGetReg(pin),
-                            1U << U_PORT_STM32F4_GPIO_PIN(pin));
+                            (uint16_t) (1U << U_PORT_STM32F4_GPIO_PIN(pin)));
 }
 
 // End of file
